requirements.txt
Django==5.1.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
PyJWT==2.9.0
asgiref==3.8.1
sqlparse==0.5.3
psycopg2-binary==2.9.10
dj-database-url==2.2.0
gunicorn==22.0.0
django-cors-headers==4.3.1

# Optional: for PostgreSQL support
# psycopg2-binary==2.9.10 